+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- `payload` command ([#21](https://github.com/gadomski/sbd-rs/pull/21))
+
 ## [0.3.1] - 2022-10-10
 
 ### Changed